Mod: Refactor code.
