+++ app/src/main/java/com/thenomads/android/nomadlive/ui/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.thenomads.android.nomadlive;
+package com.thenomads.android.nomadlive.ui;
 
 import android.annotation.TargetApi;
 import android.content.Context;
@@ -9,6 +9,8 @@ import android.preference.Preference;
 import android.preference.PreferenceActivity;
 import android.preference.PreferenceFragment;
 import android.preference.PreferenceManager;
+
+import com.thenomads.android.nomadlive.R;
 
 /**
  * A {@link PreferenceActivity} that presents a set of application settings. On
